+++ sql/backends/monet5/sql_cast_impl_int.h
@@ -42,7 +42,7 @@ FUN(do_,TP1,_dec2dec_,TP2) (TP2 *restrict res, int s1, TP1 val, int p, int s2)
 		if (val <= GDKmin(TP2) / scales[s2 - s1] ||
 		    val > GDKmax(TP2) / scales[s2 - s1]) {
 			char *buf = NULL, *msg;
-			int len = 0;
+			size_t len = 0;
 			if (BATatoms[TPE(TP1)].atomToStr(&buf, &len, &val) < 0)
 				msg = createException(SQL, "convert", SQLSTATE(22003) "value exceeds limits of type "STRNG(TP2));
 			else
@@ -55,7 +55,7 @@ FUN(do_,TP1,_dec2dec_,TP2) (TP2 *restrict res, int s1, TP1 val, int p, int s2)
 		if (val / scales[s1 - s2] <= GDKmin(TP2) ||
 		    val / scales[s1 - s2] > GDKmax(TP2)) {
 			char *buf = NULL, *msg;
-			int len = 0;
+			size_t len = 0;
 			if (BATatoms[TPE(TP1)].atomToStr(&buf, &len, &val) < 0)
 				msg = createException(SQL, "convert", SQLSTATE(22003) "Value exceeds limits of type "STRNG(TP2));
 			else
@@ -70,7 +70,7 @@ FUN(do_,TP1,_dec2dec_,TP2) (TP2 *restrict res, int s1, TP1 val, int p, int s2)
 				     ) / scales[s1 - s2]);
 	} else if (val <= GDKmin(TP2) || val > GDKmax(TP2)) {
 		char *buf = NULL, *msg;
-		int len = 0;
+		size_t len = 0;
 		if (BATatoms[TPE(TP1)].atomToStr(&buf, &len, &val) < 0)
 			msg = createException(SQL, "convert", SQLSTATE(22003) "Value exceeds limits of type "STRNG(TP2));
 		else
