+++ sql/backends/monet5/sql_statistics.c
@@ -21,30 +21,30 @@ analysis by optimizers.
 #include "sql_statistics.h"
 #include "sql_scenario.h"
 
-#define atommem(size)						\
-	do {							\
-		if (*dst == NULL || *len < (int) (size)) {	\
-			GDKfree(*dst);				\
-			*len = (size);				\
-			*dst = GDKmalloc(*len);			\
-			if (*dst == NULL)			\
-				return -1;			\
-		}						\
+#define atommem(size)					\
+	do {						\
+		if (*dst == NULL || *len < (size)) {	\
+			GDKfree(*dst);			\
+			*len = (size);			\
+			*dst = GDKmalloc(*len);		\
+			if (*dst == NULL)		\
+				return -1;		\
+		}					\
 	} while (0)
 
-static int
-strToStrSQuote(char **dst, int *len, const void *src)
+static ssize_t
+strToStrSQuote(char **dst, size_t *len, const void *src)
 {
-	int l = 0;
+	ssize_t l = 0;
 
 	if (GDK_STRNIL((str) src)) {
 		atommem(4);
 
 		return snprintf(*dst, *len, "nil");
 	} else {
-		int sz = escapedStrlen(src, NULL, NULL, '\'');
+		size_t sz = escapedStrlen(src, NULL, NULL, '\'');
 		atommem(sz + 3);
-		l = escapedStr((*dst) + 1, src, *len - 1, NULL, NULL, '\'');
+		l = (ssize_t) escapedStr((*dst) + 1, src, *len - 1, NULL, NULL, '\'');
 		l++;
 		(*dst)[0] = (*dst)[l++] = '"';
 		(*dst)[l] = 0;
@@ -62,7 +62,7 @@ sql_analyze(Client cntxt, MalBlkPtr mb, MalStkPtr stk, InstrPtr pci)
 	char *query, *dquery;
 	size_t querylen;
 	char *maxval = NULL, *minval = NULL;
-	int minlen = 0, maxlen = 0;
+	size_t minlen = 0, maxlen = 0;
 	str sch = 0, tbl = 0, col = 0;
 	int sorted, revsorted;
 	lng nils = 0;
@@ -120,7 +120,7 @@ sql_analyze(Client cntxt, MalBlkPtr mb, MalStkPtr stk, InstrPtr pci)
 						BAT *bn, *br;
 						BAT *bsample;
 						lng sz;
-						int (*tostr)(str*,int*,const void*);
+						ssize_t (*tostr)(str*,size_t*,const void*);
 						void *val=0;
 
 						if (col && strcmp(bc->name, col))
